backend/requirements.txt
aiofiles>=0.8.0
aiohttp>=3.8.0
annotated-types>=0.5.0
anyio>=3.6.0
attrs>=22.0.0
certifi>=2022.0.0
click>=8.0.0
coloredlogs>=15.0.0
google-ai-generativelanguage>=0.3.0
google-api-core>=2.11.0
google-auth>=2.17.0
google-cloud-speech>=2.18.0
google-cloud-texttospeech>=2.14.0
grpcio>=1.50.0
grpcio-status>=1.50.0
httpx>=0.24.0
livekit>=0.8.0
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.8.0
livekit-plugins-google>=0.8.0
livekit-plugins-silero>=0.8.0
numpy>=1.24.0
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
protobuf>=4.21.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
